added utils.py helper module
